+++ modules/benchpress/benchpress.dart
@@ -1,0 +1,2 @@
+export './common.dart';
+export './src/webdriver/async_webdriver_adapter.dart' show AsyncWebDriverAdapter;
