+++ modules/benchpress/benchpress.es6
@@ -1,0 +1,2 @@
+export * from './common';
+export { SeleniumWebDriverAdapter } from './src/webdriver/selenium_webdriver_adapter';
