+++ modules/benchpress/test/sampler_spec.js
@@ -8,7 +8,7 @@ import {
   Sampler, WebDriverAdapter, WebDriverExtension,
   Validator, Metric, Reporter, Browser,
   bind, Injector, Options, MeasureValues
-} from 'benchpress/benchpress';
+} from 'benchpress/common';
 
 export function main() {
   var EMPTY_EXECUTE = () => {};
